added creating dart in singleshot tower

+++ src/ooga/backend/factory/DartFactory.java
@@ -3,5 +3,5 @@ package ooga.backend.factory;
 import ooga.backend.darts.Dart;
 
 public interface DartFactory {
-  Dart createDart(int xPosition, int yPosition, int xVelocity, int yVelocity);
+  Dart createDart(int xPosition, int yPosition, double xVelocity, double yVelocity);
 }
